clean up styles and break things out into components
