+++ packages/manager/src/features/linodes/LinodesDetail/LinodeBackup/LinodeBackups.tsx
@@ -1,55 +1,43 @@
-import { LinodeBackup } from '@linode/api-v4/lib/linodes';
-import { useSnackbar } from 'notistack';
 import * as React from 'react';
-import { useHistory, useParams } from 'react-router-dom';
-import ActionsPanel from 'src/components/ActionsPanel';
+import Table from 'src/components/Table';
+import TableRowEmptyState from 'src/components/TableRowEmptyState';
 import Button from 'src/components/Button';
 import FormControl from 'src/components/core/FormControl';
-import FormHelperText from 'src/components/core/FormHelperText';
 import Paper from 'src/components/core/Paper';
-import { Theme } from '@mui/material/styles';
 import TableBody from 'src/components/core/TableBody';
 import TableHead from 'src/components/core/TableHead';
 import Typography from 'src/components/core/Typography';
-import Select from 'src/components/EnhancedSelect/Select';
 import ErrorState from 'src/components/ErrorState';
 import Notice from 'src/components/Notice';
 import TableCell from 'src/components/TableCell';
 import TableRow from 'src/components/TableRow';
 import TextField from 'src/components/TextField';
-import { resetEventsPolling } from 'src/eventsPolling';
-import { getErrorMap } from 'src/utilities/errorUtils';
-import getUserTimezone from 'src/utilities/getUserTimezone';
 import LinodePermissionsError from '../LinodePermissionsError';
 import BackupsPlaceholder from './BackupsPlaceholder';
 import BackupTableRow from './BackupTableRow';
+import { Box, Stack } from '@mui/material';
+import { Theme } from '@mui/material/styles';
+import { resetEventsPolling } from 'src/eventsPolling';
+import { getErrorMap } from 'src/utilities/errorUtils';
+import { LinodeBackup } from '@linode/api-v4/lib/linodes';
+import { useSnackbar } from 'notistack';
+import { useHistory, useParams } from 'react-router-dom';
 import { CaptureSnapshotConfirmationDialog } from './CaptureSnapshotConfirmationDialog';
 import { RestoreToLinodeDrawer } from './RestoreToLinodeDrawer';
-import {
-  useLinodeBackupSnapshotMutation,
-  useLinodeBackupsQuery,
-} from 'src/queries/linodes/backups';
-import { useFormik } from 'formik';
-import {
-  useLinodeQuery,
-  useLinodeUpdateMutation,
-} from 'src/queries/linodes/linodes';
-import { useGrants, useProfile } from 'src/queries/profile';
-import { initWindows } from 'src/utilities/initWindows';
-import Table from 'src/components/Table';
 import { useTypeQuery } from 'src/queries/types';
 import { makeStyles } from '@mui/styles';
 import { CancelBackupsDialog } from './CancelBackupsDialog';
 import { CircleProgress } from 'src/components/CircleProgress';
-import TableRowEmptyState from 'src/components/TableRowEmptyState';
+import { ScheduleSettings } from './ScheduleSettings';
+import { useGrants, useProfile } from 'src/queries/profile';
+import { useFormik } from 'formik';
+import { useLinodeQuery } from 'src/queries/linodes/linodes';
+import {
+  useLinodeBackupSnapshotMutation,
+  useLinodeBackupsQuery,
+} from 'src/queries/linodes/backups';
 
 const useStyles = makeStyles((theme: Theme) => ({
-  paper: {
-    marginBottom: theme.spacing(2),
-  },
-  subTitle: {
-    marginBottom: theme.spacing(1),
-  },
   snapshotFormControl: {
     display: 'flex',
     flexDirection: 'row',
@@ -61,20 +49,6 @@ const useStyles = makeStyles((theme: Theme) => ({
     },
     '& button': {
       marginTop: theme.spacing(4),
-    },
-  },
-  scheduleAction: {
-    padding: 0,
-    '& button': {
-      marginLeft: 0,
-      marginTop: theme.spacing(2),
-    },
-  },
-  chooseDay: {
-    marginRight: theme.spacing(2),
-    minWidth: 150,
-    '& .react-select__menu-list': {
-      maxHeight: 'none',
     },
   },
   cancelButton: {
@@ -119,12 +93,6 @@ export const LinodeBackups = () => {
   );
 
   const {
-    mutateAsync: updateLinode,
-    error: updateLinodeError,
-    isLoading: isUpdating,
-  } = useLinodeUpdateMutation(id);
-
-  const {
     mutateAsync: takeSnapshot,
     error: snapshotError,
     isLoading: isSnapshotLoading,
@@ -157,36 +125,6 @@ export const LinodeBackups = () => {
     },
   });
 
-  const settingsForm = useFormik({
-    enableReinitialize: true,
-    initialValues: {
-      day: linode?.backups.schedule.day,
-      window: linode?.backups.schedule.window,
-    },
-    async onSubmit(values) {
-      await updateLinode({
-        backups: {
-          schedule: values,
-        },
-      });
-
-      enqueueSnackbar('Backup settings saved', {
-        variant: 'success',
-      });
-    },
-  });
-
-  const days = [
-    ['Choose a day', 'Scheduling'],
-    ['Sunday', 'Sunday'],
-    ['Monday', 'Monday'],
-    ['Tuesday', 'Tuesday'],
-    ['Wednesday', 'Wednesday'],
-    ['Thursday', 'Thursday'],
-    ['Friday', 'Friday'],
-    ['Saturday', 'Saturday'],
-  ];
-
   const handleDeploy = (backup: LinodeBackup) => {
     history.push(
       '/linodes/create' +
@@ -198,18 +136,6 @@ export const LinodeBackups = () => {
     setIsRestoreDrawerOpen(true);
     setSelectedBackup(backup);
   };
-
-  const windows = initWindows(getUserTimezone(profile?.timezone), true);
-
-  const windowOptions = windows.map((window) => {
-    const label = window[0];
-    return { label, value: window[1] };
-  });
-
-  const dayOptions = days.map((day: string[]) => {
-    const label = day[0];
-    return { label, value: day[1] };
-  });
 
   const hasErrorFor = getErrorMap(['label'], snapshotError);
 
@@ -242,9 +168,9 @@ export const LinodeBackups = () => {
       Boolean(backups?.snapshot.in_progress));
 
   return (
-    <React.Fragment>
+    <Stack spacing={2}>
       {doesNotHavePermission && <LinodePermissionsError />}
-      <Paper className={classes.paper} style={{ padding: 0 }}>
+      <Paper style={{ padding: 0 }}>
         <Table aria-label="List of Backups">
           <TableHead>
             <TableRow>
@@ -303,12 +229,8 @@ export const LinodeBackups = () => {
           </TableBody>
         </Table>
       </Paper>
-      <Paper className={classes.paper}>
-        <Typography
-          variant="h2"
-          className={classes.subTitle}
-          data-qa-manual-heading
-        >
+      <Paper>
+        <Typography variant="h2" data-qa-manual-heading>
           Manual Snapshot
         </Typography>
         <Typography variant="body1" data-qa-manual-desc>
@@ -342,107 +264,32 @@ export const LinodeBackups = () => {
           </Button>
         </FormControl>
       </Paper>
+      <ScheduleSettings linodeId={id} />
+      <Box>
+        <Button
+          buttonType="outlined"
+          className={classes.cancelButton}
+          disabled={doesNotHavePermission}
+          onClick={() => setIsCancelBackupsDialogOpen(true)}
+          data-qa-cancel
+        >
+          Cancel Backups
+        </Button>
+        <Typography
+          className={classes.cancelCopy}
+          variant="body1"
+          data-qa-cancel-desc
+        >
+          Please note that when you cancel backups associated with this Linode,
+          this will remove all existing backups.
+        </Typography>
+      </Box>
       <CaptureSnapshotConfirmationDialog
         open={isSnapshotConfirmationDialogOpen}
         onClose={() => setIsSnapshotConfirmationDialogOpen(false)}
         onSnapshot={() => snapshotForm.handleSubmit()}
         loading={isSnapshotLoading}
       />
-      <Paper className={classes.paper}>
-        <form onSubmit={settingsForm.handleSubmit}>
-          <Typography
-            variant="h2"
-            className={classes.subTitle}
-            data-qa-settings-heading
-          >
-            Settings
-          </Typography>
-          <Typography variant="body1" data-qa-settings-desc>
-            Configure when automatic backups are initiated. The Linode Backup
-            Service will generate a backup between the selected hours every day,
-            and will overwrite the previous daily backup. The selected day is
-            when the backup is promoted to the weekly slot. Up to two weekly
-            backups are saved.
-          </Typography>
-          {Boolean(updateLinodeError) && (
-            <Notice error spacingTop={16} spacingBottom={0}>
-              {updateLinodeError?.[0].reason}
-            </Notice>
-          )}
-          <FormControl className={classes.chooseDay}>
-            <Select
-              textFieldProps={{
-                dataAttrs: {
-                  'data-qa-weekday-select': true,
-                },
-              }}
-              options={dayOptions}
-              onChange={(item) => settingsForm.setFieldValue('day', item.value)}
-              value={dayOptions.find(
-                (item) => item.value === settingsForm.values.day
-              )}
-              label="Day of Week"
-              placeholder="Choose a day"
-              isClearable={false}
-              name="Day of Week"
-              noMarginTop
-            />
-          </FormControl>
-          <FormControl>
-            <Select
-              textFieldProps={{
-                dataAttrs: {
-                  'data-qa-time-select': true,
-                },
-              }}
-              options={windowOptions}
-              onChange={(item) =>
-                settingsForm.setFieldValue('window', item.value)
-              }
-              value={windowOptions.find(
-                (item) => item.value === settingsForm.values.window
-              )}
-              label="Time of Day"
-              placeholder="Choose a time"
-              isClearable={false}
-              name="Time of Day"
-              noMarginTop
-            />
-            <FormHelperText sx={{ marginLeft: 0 }}>
-              Time displayed in{' '}
-              {getUserTimezone(profile?.timezone).replace('_', ' ')}
-            </FormHelperText>
-          </FormControl>
-          <ActionsPanel className={classes.scheduleAction}>
-            <Button
-              buttonType="primary"
-              type="submit"
-              disabled={doesNotHavePermission || !settingsForm.dirty}
-              loading={isUpdating}
-              data-qa-schedule
-            >
-              Save Schedule
-            </Button>
-          </ActionsPanel>
-        </form>
-      </Paper>
-      <Button
-        buttonType="outlined"
-        className={classes.cancelButton}
-        disabled={doesNotHavePermission}
-        onClick={() => setIsCancelBackupsDialogOpen(true)}
-        data-qa-cancel
-      >
-        Cancel Backups
-      </Button>
-      <Typography
-        className={classes.cancelCopy}
-        variant="body1"
-        data-qa-cancel-desc
-      >
-        Please note that when you cancel backups associated with this Linode,
-        this will remove all existing backups.
-      </Typography>
       <RestoreToLinodeDrawer
         open={isRestoreDrawerOpen}
         linodeId={id}
@@ -454,7 +301,7 @@ export const LinodeBackups = () => {
         onClose={() => setIsCancelBackupsDialogOpen(false)}
         linodeId={id}
       />
-    </React.Fragment>
+    </Stack>
   );
 };
